It sticks to the original PoW/PoS/Treasury split of 60/30/10, a much more reasonable and healthier balance.
A DCRN forking is inevitable to do to keep the essence and advantage of the original and the authentic DCR.

We all know that PoS is a ‘winners take all’ game where the whales’ powers snowball over time, and becoming the de-facto centralised controller of an ecosystem, and this is not just some daunting imagination, it is a widely existing problem for public blockchains that run on the PoS consensus. What’s worse, it has already been proved after DCR’s subsidy split change: total amount of staking remained the same, but subsidy was taken from part of the decreased PoW miners and re-allocated to early PoS pledger. The result? Devastating and unfair to both early-stage and new-joining PoW miners.

In reality, we are heading towards a world where PoW and PoS co-exist, it’s never an ‘either or’ game, as that’s never healthy. We can all agree that both PoW and PoS possess each of their own advantages and disadvantages. The drastic increase in the PoS split is a catastrophe to the PoW miners and given how significant a part they played in the classic DCR community, it’s also a catastrophe to the DCR world by exploiting these miners.